requirements.txt
Django>=5.2,<5.3
gunicorn>=22.0,<23.0
psycopg[binary]>=3.2,<4.0
djangorestframework>=3.15,<4.0
djangorestframework-gis>=1.1,<2.0
django-filter>=24.0,<25.0
Pillow>=11.0,<12.0
requests>=2.32,<3.0
geopy>=2.4,<3.0
python-decouple>=3.8,<4.0
